MPT-16437 Reorganise resource catalog mixins structure
